Make name space for optimizer hooks and move optimizer hooks there
