+++ chainer/optimizer_hooks/gradient_noise.py
@@ -1,0 +1,51 @@
+from chainer import cuda
+
+
+class GradientNoise(object):
+    """Optimizer hook function for adding gradient noise.
+
+    This hook function simply adds noise generated by the ``noise_func``
+    to the gradient. By default it adds time-dependent annealed Gaussian
+    noise to the gradient at every training step:
+
+    .. math::
+
+        g_t \\leftarrow g_t + N(0, \\sigma_t^2)
+
+    where
+
+    .. math::
+
+        \\sigma_t^2 = \\frac{\\eta}{(1+t)^\\gamma}
+
+    with :math:`\\eta` selected from {0.01, 0.3, 1.0} and
+    :math:`\\gamma = 0.55`.
+
+    Args:
+        eta (float): Parameter that defines the scale of the noise, which for
+            the default noise function is recommended to be either 0.01, 0.3
+            or 1.0.
+        noise_func (function): Noise generating function which by default
+            is given by `Adding Gradient Noise Improves Learning for Very Deep\
+            Networks <https://arxiv.org/pdf/1511.06807>`_.
+    """
+    name = 'GradientNoise'
+
+    def __init__(self, eta, noise_func=exponential_decay_noise):
+        self.eta = eta
+        self.noise_func = noise_func
+
+    def kernel(self):
+        return cuda.elementwise(
+            'T noise', 'T g', 'g += noise', 'gradient_noise')
+
+    def __call__(self, opt):
+        for param in opt.target.params():
+            g = param.grad
+            xp = cuda.get_array_module(g)
+            with cuda.get_device(g) as dev:
+                noise = self.noise_func(xp, g.shape, g.dtype, self, opt)
+                if int(dev) == -1:
+                    g += noise
+                else:
+                    self.kernel()(noise, g)
