+++ docs/source/reference/core/optimizer.rst
@@ -7,9 +7,3 @@ Optimizer
 .. autoclass:: GradientMethod
    :members:
 
-Hook functions
-~~~~~~~~~~~~~~
-.. autoclass:: chainer.optimizer.WeightDecay
-.. autoclass:: chainer.optimizer.Lasso
-.. autoclass:: chainer.optimizer.GradientClipping
-.. autoclass:: chainer.optimizer.GradientNoise
