+++ docs/source/reference/optimizer_hooks.rst
@@ -1,0 +1,7 @@
+Hook functions
+~~~~~~~~~~~~~~
+.. autoclass:: chainer.optimizers.WeightDecay
+.. autoclass:: chainer.optimizers.Lasso
+.. autoclass:: chainer.optimizers.GradientClipping
+.. autoclass:: chainer.optimizers.GradientHardClipping
+.. autoclass:: chainer.optimizers.GradientNoise
