+++ examples/cifar/train_cifar.py
@@ -63,7 +63,7 @@ def main():
 
     optimizer = chainer.optimizers.MomentumSGD(0.1)
     optimizer.setup(model)
-    optimizer.add_hook(chainer.optimizer.WeightDecay(5e-4))
+    optimizer.add_hook(chainer.optimizers.WeightDecay(5e-4))
 
     train_iter = chainer.iterators.SerialIterator(train, args.batchsize)
     test_iter = chainer.iterators.SerialIterator(test, args.batchsize,
