+++ setup.py
@@ -55,6 +55,7 @@ setup(
               'chainer.links.normalization',
               'chainer.links.theano',
               'chainer.optimizers',
+              'chainer.optimizer_hooks',
               'chainer.serializers',
               'chainer.testing',
               'chainer.training',
